Update README.md
In this project:
I made use of Python to explore data related to bike share systems for three major cities in the United States—Chicago, New York City, and Washington. 
I wrote code to import the data and answer interesting questions about it by computing descriptive statistics. 
I also wrote a script that takes in raw input to create an interactive experience in the terminal to present these statistics.
The BikeShare-FWD-Udacity.py file is set up as a script that takes in raw input to create an interactive experience in the terminal that answers questions about the dataset. The experience is interactive because depending on a user's input, the answers to the questions on the previous page will change! There are four questions that will change the answers:
Would you like to see data for Chicago, New York, or Washington?
Would you like to filter the data by month, day, or not at all?
(If they chose month) Which month - January, February, March, April, May, or June?
(If they chose day) Which day - Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, or Sunday?
The answers to the questions above will determine the city and timeframe on which we can do data analysis. 
After filtering the dataset, users will see the statistical result of the data, and choose to start again or exit.